requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
email-validator==2.0.0.post2
activecampaign-api-python==1.0.2
systeme-api==0.1.0
python-mimeparse==1.6.0
redis==5.0.1
python-dateutil==2.8.2
SQLAlchemy==2.0.20
psycopg2-binary==2.9.9
alembic==1.12.1
